+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Changed
+- Field layout tabs for new Neo input blocks of a certain block type are now lazy loaded when a block of that type is first created
+
 ### Fixed
 - Fixed a bug where Neo block structure data would be lost when soft-deleting an entry
 - Fixed an error that could occur when applying Neo block type project config changes